Move list page displays to under components/ListPageDisplays
